requirements.txt
Flask==2.0.1
pymongo==3.12.0
Pillow==8.3.1